Module Elaboration.Var

module B = Kernel.Basic

module F = Common.Files

module T = Kernel.Term

module U = Common.Universes

exception Not_uvar

val basename : string

prefix for all the universe variables.

val is_pre_var : T.term -> bool

Check if a term should be elaborated by a fresh variable

val is_uvar : T.term -> bool

Check if a term is universe variable, i.e. its ident should be ?11, ?43...

val name_of_uvar : T.term -> Kernel.Basic.name

name_of_uvar t returns the name of universe variable if t is a universe variable, raise Not_uvar otherwise

val counter : int Stdlib.ref

Internal counter use by this module to generate fresh variables

val fresh : unit -> B.ident

Generate a fresh name for a universe variable

val fresh_uvar : F.cout F.t -> unit -> T.term

fresh_uvar env () returns a fresh term representing a universe variable. Add a new declaration into the module env.md